refactor: extracted keyboard navigation
- now the keyboard navigation is potentially reusable and can be also exposed as standalone hook if needed
